modified commit

+++ src/algorithms/Search.java
@@ -4,7 +4,7 @@ package algorithms;
 
 public class Search {
 
-	public static double dataSearch(int[] numbers, double target) {
+	public static double dataSearch(int[] numbers, int target) {
 		
 		for(int i =0;i<numbers.length;i++) {
 		   if(numbers[i]==target)
